// Tests/UI/SplittingClient/src/api/models/Order.ts
//----------------------
// <auto-generated>
//     Generated using the NSwag toolchain v14.0.8.0 (NJsonSchema v11.0.1.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
// </auto-generated>
//----------------------

import { OrderItem } from './OrderItem';
import { OrderStatus } from './OrderStatus';

export interface Order {
    id: string;
    userId: string;
    items: OrderItem[];
    total: number;
    status: OrderStatus;
    createdAt: Date;

    [key: string]: any;
}