// Tests/UI/SplittingClient/src/api/models/User.ts
//----------------------
// <auto-generated>
//     Generated using the NSwag toolchain v14.0.8.0 (NJsonSchema v11.0.1.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
// </auto-generated>
//----------------------

import { UserRole } from './UserRole';

export interface User {
    id: string;
    username: string;
    email: string;
    role: UserRole;
    createdAt: Date;

    [key: string]: any;
}