// Tests/UI/SplittingClient/src/api/models/Good.ts
//----------------------
// <auto-generated>
//     Generated using the NSwag toolchain v14.0.8.0 (NJsonSchema v11.0.1.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
// </auto-generated>
//----------------------

export interface Good {
    id: string;
    name: string;
    description: string;
    price: number;
    stock: number;
    createdAt: Date;

    [key: string]: any;
}